requirements.txt
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.63
langchain-text-splitters==0.3.8
langchain-groq==0.2.3
langsmith==0.1.147
faiss-cpu==1.13.2
pypdf==4.3.1
python-dotenv==1.0.1
streamlit==1.45.0
sentence-transformers==3.3.1
groq==0.9.0
tokenizers==0.21.1
transformers==4.51.0
httpx==0.27.2
pydantic==2.10.6